refactor: randomUtil 설정
